requirements.txt
groq==0.4.1
elevenlabs==0.2.26
fal-client==0.4.0
openai==1.13.3
librosa==0.10.1
google-cloud-storage==2.10.0
requests==2.31.0
python-dotenv==1.0.0
ffmpeg-python==0.2.0
pydub==0.25.1
scipy==1.11.4
numpy==1.24.3
pillow==10.0.1
aiohttp==3.9.1
asyncio
uuid
json
logging
dataclasses
typing
pathlib